moved stuff from util
