+++ pkg/dockerclient/dockerclient.go
@@ -21,8 +21,8 @@ import (
 	"path"
 	"strings"
 
+	"github.com/nuclio/nuclio/pkg/cmdrunner"
 	"github.com/nuclio/nuclio/pkg/errors"
-	"github.com/nuclio/nuclio/pkg/util/cmdrunner"
 
 	"github.com/nuclio/nuclio-sdk"
 )
