+++ pkg/processor/eventsource/kafka/eventsource.go
@@ -17,10 +17,10 @@ limitations under the License.
 package kafka
 
 import (
+	"github.com/nuclio/nuclio/pkg/common"
 	"github.com/nuclio/nuclio/pkg/errors"
 	"github.com/nuclio/nuclio/pkg/processor/eventsource"
 	"github.com/nuclio/nuclio/pkg/processor/worker"
-	"github.com/nuclio/nuclio/pkg/util/common"
 
 	"github.com/Shopify/sarama"
 	"github.com/nuclio/nuclio-sdk"
