+++ pkg/processor/eventsource/kinesis/eventsource.go
@@ -17,10 +17,10 @@ limitations under the License.
 package kinesis
 
 import (
+	"github.com/nuclio/nuclio/pkg/common"
 	"github.com/nuclio/nuclio/pkg/errors"
 	"github.com/nuclio/nuclio/pkg/processor/eventsource"
 	"github.com/nuclio/nuclio/pkg/processor/worker"
-	"github.com/nuclio/nuclio/pkg/util/common"
 
 	"github.com/nuclio/nuclio-sdk"
 	kinesisclient "github.com/sendgridlabs/go-kinesis"
