+++ pkg/processor/eventsource/poller/v3ioitempoller/eventsource.go
@@ -21,10 +21,10 @@ import (
 	"strings"
 	"sync"
 
+	"github.com/nuclio/nuclio/pkg/common"
 	"github.com/nuclio/nuclio/pkg/processor/eventsource"
 	"github.com/nuclio/nuclio/pkg/processor/eventsource/poller"
 	"github.com/nuclio/nuclio/pkg/processor/worker"
-	"github.com/nuclio/nuclio/pkg/util/common"
 
 	"github.com/iguazio/v3io-go-http"
 	"github.com/nuclio/nuclio-sdk"
