+++ pkg/registry/registry_test.go
