+++ pkg/restful/resource.go
@@ -19,7 +19,7 @@ package restful
 import (
 	"net/http"
 
-	"github.com/nuclio/nuclio/pkg/util/registry"
+	"github.com/nuclio/nuclio/pkg/registry"
 
 	"github.com/go-chi/chi"
 	"github.com/nuclio/nuclio-sdk"
